+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 dev
 ---
 
-* -
+* Fixed the use of environment variable ``RETDEC_API_URL`` to override the
+  default API URL. Previously, the URL was not overridden even when given.
 
 0.5 (2016-09-15)
 ----------------