+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ dev
   downloading all the requested outputs instead of giving up on the first
   failure.
 * Fixed the use of environment variable ``RETDEC_API_URL`` to override the
-  default API URL. Previously, the URL was not overridden even when given.
+  default API URL. Previously, the URL was not overridden, even when given.
 
 0.5 (2016-09-15)
 ----------------